+++ fabric/src/memory/global.hpp
@@ -28,11 +28,11 @@ namespace veil::memory {
     /// allocation.
     class HeapObject {
     public:
-        void *operator new(uint64 size);
+        void *operator new(size_t size);
 
         void operator delete(void *address);
 
-        void *operator new[](uint64 size) = delete;
+        void *operator new[](size_t size) = delete;
 
         void operator delete[](void *address) = delete;
     };
@@ -41,11 +41,11 @@ namespace veil::memory {
     /// this class, this class forbids descendants to be allocated to the process heap.
     class ValueObject {
     public:
-        void *operator new(uint64 size) = delete;
+        void *operator new(size_t size) = delete;
 
         void operator delete(void *address) = delete;
 
-        void *operator new[](uint64 size) = delete;
+        void *operator new[](size_t size) = delete;
 
         void operator delete[](void *address) = delete;
     };
